1. ModuleBsin - Implemented Eduardo straightforward code for managing other modules dt's. Added that if atmospheredt is zero (happens if method for predicting atmosphere dt's is the old, and means that current time is the same as one HDF precipitation instant) than this value is not used because has no valid info and would made the model to crash. 2. ModuleAtmosphere - do not let other properties than precipitation and irrigation to have accumulated values. 3. In other modules some small changes so that the code compiles in single configurations.
git-svn-id: https://mohid.svn.codeplex.com/svn@87566 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/MOHIDLand/ModulePorousMediaProperties.F90
@@ -701,7 +701,7 @@ Module ModulePorousMediaProperties
         
         real(8), pointer, dimension(:,:,:)           :: WaterVolume     => null()
         real(8), pointer, dimension(:,:,:)           :: FluxWCorr       => null()
-        real(8), pointer, dimension(:,:,:)           :: WaterContentBT  => null()  !Water Content Before Transport
+        real,    pointer, dimension(:,:,:)           :: WaterContentBT  => null()  !Water Content Before Transport
         
         logical                                      :: DTIntervalAssociated     = .false.
 
@@ -5492,7 +5492,7 @@ cd1 :   if ((ready_ .EQ. IDLE_ERR_     ) .OR. &
         real, dimension(:, :   ), pointer       :: DischargesConc
         type (T_Property), pointer              :: Property
         integer                                 :: iProp
-        real                                    :: VolumeOld
+        real(8)                                 :: VolumeOld
 
 
         !Get integrated flow from runoff to be sure using same values
@@ -8322,7 +8322,7 @@ doi4 :      do i = Me%WorkSize%ILB, Me%WorkSize%IUB
         
         !Local--------------------------------------------------------------------
         real, pointer, dimension(:,:,:)             :: TranspirationLayer
-        real, pointer, dimension(:,:  )             :: Evaporation
+        real(8), pointer, dimension(:,:  )          :: Evaporation
         logical                                     :: TranspirationExists, EvaporationExists
         integer                                     :: i,j,k, CHUNK, STAT_CALL
         real(8)                                     :: aux
